rename cases.py to cases_cb.py
